+++ tests/args.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+
+class Args:
+    def __init__(self):
+        self.seed = "1"
+        self.left_to_right = True
+        self.right_to_left = False
+        self.top_down = False
+        self.bottom_up = False
+        self.zig_zag = False
+        self.zig_zag_rev = False
+        self.minlen = 6
+        self.maxlen = 8